Make the closing of the tab a bit smarter, v1.1
The tab was always being closed before, which was a problem if you for
instance were reading a blog post and clicked the link inside there.

Now the extension tries to be a bit smarter about this and only closes
the tab if it is newly opened, or as close as I've been able to figure out.